function

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	int i = 1, j = 2, n, x = 0;
+	long int i = 1, j = 2, n, x = 0;
 
 	printf("%d, %d", i, j);
 	for (n = 0; n < 48; n++)
